requirements.txt
Flask==2.3.3
cohere==4.27
python-dotenv==1.0.0
requests==2.31.0
